+++ main.py
@@ -26,48 +26,17 @@ from kivy.uix.button import Button
 from kivy.uix.label import Label
 from kivy.uix.popup import Popup
 from kivy.uix.screenmanager import ScreenManager, Screen
-from kivy.storage.jsonstore import JsonStore
 from kivy.uix.textinput import TextInput
 from os.path import join
 import re
 
+from store import get_store, create_store
 
-__version__ = '0.0.1'
+
+__version__ = '0.0.2'
 
 
-def get_store():
-    store = JsonStore('cetoolboxdata.json')
-    return store
-
-def create_store():
-    """ function to create a file with default value of each var"""
-    store = get_store()
-    if not store.exists('Capillary'):
-        store.put('Capillary', value=100.00, unit="cm")
-    if not store.exists('Towindow'):
-        store.put('Towindow', value=100.00, unit="cm")
-    if not store.exists('Idiameter'):
-        store.put('Idiameter', value=50.00, unit="µm")
-    if not store.exists('Pressure'):
-        store.put('Pressure', value=50.00, unit="mbar")
-    if not store.exists('Time'):
-        store.put('Time', value=50.00, unit="s")
-    if not store.exists('Viscosity'):
-        store.put('Viscosity', value=1.0, unit="cp")
-    if not store.exists('Concentration'):
-        store.put('Concentration', value=1.0, unit="g/L")
-    if not store.exists('Molweight'):
-        store.put('Molweight', value=1000.0, unit="g/mol")
-    if not store.exists('Detectiontime'):
-        store.put('Detectiontime', value=10.0, unit="m")
-    if not store.exists('Voltage'):
-        store.put('Voltage', value=3.0, unit="kV")
-    if not store.exists('Electriccurrent'):
-        store.put('Electriccurrent', value=10.0, unit="µA")
-    if not store.exists('Electroosmosis'):
-        store.put('Electroosmosis', value=1.0, unit="m")
     
-
 class FloatInput(TextInput):
     pat = re.compile('[^0-9]')
     def insert_text(self, substring, from_undo=False):
